Renamed using "assertion" instead of "triple"
Instead of referencing opposite triples, we now reference assertions of oppositeness. This resulted in changes to code, scripts, output files, Dockerfiles, etc.
