+++ classes/scripts/metakg-assertions-of-oppositeness.entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+mvn exec:java -Dexec.mainClass='edu.cuanschutz.ccp.opposites.MetaKgOppositeAssertionExtractorMain' -Dexec.args='/home/dev/data/predicates.txt /home/dev/output'
