--- classes/scripts/metakg-opposite-triple.entrypoint.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-mvn exec:java -Dexec.mainClass='edu.cuanschutz.ccp.opposites.MetaKgOppositeTripleExtractorMain' -Dexec.args='/home/dev/data/predicates.txt /home/dev/output'
